Cleaning up torchwm
