+++ .github/workflows/docs.yml
@@ -10,6 +10,7 @@ on:
     paths:
       - "docs/**"
       - "world_models/**"
+      - "torchwm/**"
       - ".github/workflows/docs.yml"
   workflow_dispatch:
 
@@ -21,8 +22,13 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  build:
+  docs-check:
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ['3.11', '3.12', '3.13']
+
     steps:
       - name: Checkout
         uses: actions/checkout@v4
@@ -30,14 +36,49 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.13"
+          python-version: ${{ matrix.python-version }}
 
       - name: Cache pip
         uses: actions/cache@v4
         with:
           path: ~/.cache/pip
-          key: ${{ runner.os }}-pip-docs-${{ hashFiles('docs/source/**', 'pyproject.toml', '.github/workflows/docs.yml') }}
+          key: ${{ runner.os }}-py${{ matrix.python-version }}-pip-docs-${{ hashFiles('docs/source/**', 'pyproject.toml', '.github/workflows/docs.yml') }}
           restore-keys: |
+            ${{ runner.os }}-py${{ matrix.python-version }}-pip-docs-
+            ${{ runner.os }}-pip-docs-
+
+      - name: Install documentation dependencies
+        run: |
+          python -m pip install --upgrade pip
+          sudo apt-get update && sudo apt-get install -y pandoc
+          pip install sphinx myst-parser sphinx-rtd-theme sphinx-autodoc-typehints sphinxext.opengraph pydata-sphinx-theme sphinx-copybutton sphinxcontrib-mermaid sphinx-thebe nbsphinx matplotlib attrdict
+          pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
+          pip install numpy gym gymnasium tensorboardX tensorboard plotly pillow tqdm einops pyyaml pygame opencv-python albumentations cma
+
+      - name: Build HTML docs
+        env:
+          SPHINX_HTML_BASEURL: ${{ github.event_name == 'pull_request' && format('https://paramthakkar123.github.io/torchwm/pr-preview/pr-{0}/', github.event.number) || 'https://paramthakkar123.github.io/torchwm/' }}
+        run: sphinx-build -b html docs/source docs/build/html
+
+  build:
+    runs-on: ubuntu-latest
+    needs: docs-check
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.13'
+
+      - name: Cache pip
+        uses: actions/cache@v4
+        with:
+          path: ~/.cache/pip
+          key: ${{ runner.os }}-pip-docs-deploy-${{ hashFiles('docs/source/**', 'pyproject.toml', '.github/workflows/docs.yml') }}
+          restore-keys: |
+            ${{ runner.os }}-pip-docs-deploy-
             ${{ runner.os }}-pip-docs-
 
       - name: Install documentation dependencies
