+++ .github/workflows/publish-pypi.yml
@@ -54,7 +54,7 @@ jobs:
           print(f"version={version}")
           PY
 
-      - name: Update pyproject.toml version
+      - name: Update version source
         env:
           PACKAGE_VERSION: ${{ steps.release-version.outputs.version }}
         run: |
@@ -64,16 +64,16 @@ jobs:
           from pathlib import Path
 
           version = os.environ["PACKAGE_VERSION"]
-          path = Path("pyproject.toml")
+          path = Path("world_models/_version.py")
           text = path.read_text(encoding="utf-8")
           updated = re.sub(
-              r'(?m)^(version\s*=\s*)["\'][^"\']+["\']',
+              r'(?m)^(__version__\s*=\s*)["\'][^"\']+["\']',
               rf'\1"{version}"',
               text,
               count=1,
           )
           if updated == text:
-              raise SystemExit("Could not find a top-level project version in pyproject.toml")
+              raise SystemExit("Could not find __version__ in world_models/_version.py")
           path.write_text(updated, encoding="utf-8")
           PY
 
@@ -86,9 +86,9 @@ jobs:
           git config user.name "github-actions[bot]"
           git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
 
-          git add pyproject.toml
+          git add world_models/_version.py
           if git diff --cached --quiet; then
-            echo "pyproject.toml already contains version ${PACKAGE_VERSION}."
+            echo "world_models/_version.py already contains version ${PACKAGE_VERSION}."
             exit 0
           fi
