+++ pyproject.toml
@@ -17,11 +17,10 @@ classifiers = [
 dependencies = [
     "torch>=2.1.0",
     "torchvision>=0.16.0",
+    "click>=8.0.0",
     "einops>=0.8.2",
-    "gymnasium[box2d]>=1.2.2",
     "tqdm>=4.67.1",
     "pyyaml>=6.0.3",
-    "wandb>=0.16.0",
 ]
 
 [project.urls]
@@ -31,11 +30,12 @@ Issues = "https://github.com/paramthakkar123/torchwm/issues"
 Changelog = "https://github.com/paramthakkar123/torchwm/releases"
 
 [project.scripts]
-torchwm = "tools.cli:run"
-torchwm-train = "tools.cli:run_train"
+torchwm = "torchwm.cli:run"
+torchwm-train = "torchwm.cli:run_train"
 
 [project.optional-dependencies]
 gym = [
+    "gymnasium[box2d]>=1.2.2",
     "huggingface-hub>=0.23.0",
     "pygame>=2.6.1",
     "autorom>=0.6.1",
@@ -73,9 +73,9 @@ viz = [
 ]
 ml = [
     "torchaudio>=0.13.0",
+    "wandb>=0.16.0",
     "tensorboard>=2.20.0",
     "tensorboardx>=2.6.4",
-    "wandb>=0.16.0",
 ]
 dev = [
     "pytest>=9.0.3",
@@ -108,7 +108,7 @@ docs = [
 ]
 
 [tool.setuptools.packages.find]
-include = ["world_models", "world_models.*", "torchwm", "torchwm.*"]
+include = ["world_models", "world_models.*", "torchwm", "torchwm.*", "tools", "tools.*"]
 
 [tool.setuptools.package-data]
 torchwm = ["py.typed", "__init__.pyi"]
@@ -137,3 +137,4 @@ markers = [
     "integration: tests that exercise external services, optional backends, or cross-package workflows",
     "selenium: tests that require Selenium WebDriver and a running web stack",
 ]
+
