+++ setup.py
@@ -3,6 +3,6 @@
 All canonical metadata lives in ``pyproject.toml``.
 """
 
-from setuptools import setup, find_namespace_packages
+from setuptools import setup
 
-setup(packages=find_namespace_packages(include=["world_models*", "torchwm*"]))
+setup()
