+++ tools/__init__.py
@@ -1,6 +1,6 @@
 """Tooling helpers for TorchWM.
 
-All heavyweight or optional helpers are exposed lazily so ``python -m tools.cli``
+All heavyweight or optional helpers are exposed lazily so ``python -m torchwm.cli``
 and the installed ``torchwm`` command do not import documentation/browser tooling
 or pre-import the CLI module during package initialization.
 """
@@ -58,3 +58,4 @@ def __getattr__(name: str) -> Any:
 
     globals()[name] = value
     return value
+
